ACCUMULO-1605
Added many finally clauses to close spans in the presence of exceptions.

I suspect that the extreme depth of the traces is caused by traces that
are started and never stopped.  This most likely happens in the presence
of errors, where exceptions jump around the the span.stop calls.
In particular, when a file is missing for a compaction.  This is known
to occur in 1.4 when bulk imports fail.  See ACCUMULO-1044.