ACCUMULO-1158 Update Eclipse formatter
The Eclipse formatter change in ACCUMULO-1158 was reverted along with
all the changes to existing files, to prevent indenting blank lines,
because this caused a problem with merges in subversion. I don't
anticipate that being a problem in git, but just in case, I won't make
the change on all existing files. People can transition to this
formatter over time.