ACCUMULO-1773 Review and minor changes to GarbageCollection rework.
Add shortcircuit when scanning candidates for any files which might be in a
bulk load. Add javadoc to the GarbageCollectionEnvironment to actually describe
what the interface should be doing. Remove old 'offline' warning message
as the option itself is gone. Added TODO about caught exception when
trying to write deletes. Removed unnecessary string concat in
ChooseDir.