ACCUMULO-1829 Add timeout scaling factor to tests (1.4.x)
Slower hardware or VMs may have trouble completing tests within the timeouts
specified. This change adds a -f option to run.py to add a scaling factor
to the maxRuntime for tests, the timeout for Accumulo to shut down, and to
specific timeouts in SimpleBalancerFairness and ZooCacheTest.

Compare to a similar change for ACCUMULO-1789.

Signed-off-by: Eric Newton <eric.newton@gmail.com>