Revert "ACCUMULO-1030 - accumulo-maven-plugin skipped with skipTests"
This reverts commit 9d6da9698d7a8ae35fa72ce0b48caa0e8124c497.

This commit was done to fix ACCUMULO-1675, but does not correctly do
that. ACCUMULO-1675 was supposed to skip execution of the plugin on a
project that uses it. It was *NOT* supposed to offer a way to avoid
building the plugin itself. Even if we skip tests in the Accumulo build,
we *still* need to build/package/install/deploy this plugin.

To correctly fix ACCUMULO-1675, we need to modify the plugin itself to
behave differently upon encountering -DskipTests (or another similar
flag) on a project that uses this plugin for integration testing against
Accumulo.