ACCUMULO-2027 Synchronized access to ZooKeeperInstance methods that mutated state
The usage pattern of a ZooKeeperInstance is to have many clients use the same
instance. Internall, the ZooKeeperInstances keeps a cache of information in a
ZooCache, which will manage connections to Zookeeper for getting and putting
data.

With the addition of close semantics on the ZKI, it is expected the ZKI will
free resources and halt operations for all clients. This requires ZKI method
calls to be linearizable, otherwise race conditions as described in
ACCUMULO-2027 will occur when a ZKI is simultaneously asked to close and
retrieve information.

This patch resolves that race condition by synchronizing all methods on a ZKI
that interact with an ZKI instance's ZooCache.