ACCUMULO-2102 Fix for intermittent failures of ShellSetInstanceTest by Bill Havanki
I think the problem may lie in the use of SiteConfiguration by the Shell. It only creates a single static
instance object during the run, so after the first time it is used in a unit test, subsequent unit tests
should receive the same object with an old mock inside, instead of a new one with a new mock.