ACCUMULO-2170 Add in a tag element to the scm block
The release plugin will add in the tag elemen when it's not
present, by appending it to the last child in the scm element.
This creates a non-sorted pom, which sortpom will fail the build over.
Adding in a tag with a "stub" value will at least let the build
pass with the release plugin.