ACCUMULO-2334 Remove ACCUMULO_LOG_HOST in favor of pull host and port log-forwarding from zookeeper
Advertising both the host and port for log4j gives us a couple of benefits.
We can do away with ACCUMULO_LOG_HOST, simplify the code to always
do the same thing (pull from zookeeper), and gain robust failover
if the monitor is moved to a different host or is restarted with a random
port (does not require any other service restart to become aware).

The monitor will now acquire a zoolock before starting, which ensures that
all tservers will perform log-forwarding to the correct monitor (in the case
that there were multiple for some reason). Creates a better hierarchy in ZooKeeper
for all monitor data (HTTP and Log4j advertisement). Update `accumulo init`,
`accumulo info` and ensure that the zookeeper layout can handle an "upgrade"
from 1.5.0.