ACCUMULO-2488 Change criteria for unbalanced servers in concurrent randomwalk
The Concurrent randomwalk test used to consider servers unbalanced if any server's
tablet count differed from the cluster average by more than a fifth of the average or
by one, whichever was larger. This would cause failures under typical balancings from
the default balancer.

This commit changes the criterion for an unbalanced server to be double the standard
deviation from the cluster average.