ACCUMULO-2621 Add delay to Shutdown step of continuous randomwalk, update README
The Concurrent.xml randomwalk module can shut down and restart the Accumulo cluster during
the test. This commit adds a 10 second delay after shutdown, to give the master plenty of
time to exit before the test attempts to start it up again.

Also, it is essential that the walker(s) running the test have passwordless SSH access to
the cluster to run $ACCUMULO_HOME/bin/start-all.sh. Otherwise, the masters will not restart
and the test will eventually get stuck. The randomwalk README.md now has information about
that.