ACCUMULO-2212 Add ZooReaderWriterFactory
The o.a.a.server.zookeeper.ZooReaderWriter class is mostly copied into a new
ZooReaderWriterFactory class. Users of the factory can work directly with FATE
ZooReaderWriter instances, instead of those of the server type, which is no longer
needed. The logic for building instances from site configuration data is in the
factory.

As part of this change, the invocation handler used to retry ZK calls on connection
loss is refactored into a new RetryingInvocationHandler.