ACCUMULO-2665 Workaround for hadoop-2.4.0 requiring a <Guava-15.0
Guava 15.0 removes the LimitInputStream class that hdfs uses internally. We need
to ensure that we have such a version on the classpath for tests that use MiniDFSCluster.
The test module is not prone to this due to sisu-guava being pulled in from maven -- avoiding
fixing that as well since this is a very ugly fix. This is a bandaid in term for a better long-term fix.