ACCUMULO-2702 Create separate MR Client Opts
Separate the MR opts from the core client opts. There is plenty more
work to be done, but this is a reasonable first step in that direction.

We should probably be using a lot more composition than inheritance
overall, and this resulted in some slight code duplication, but the
change set would have been *much* more extensive otherwise.