ACCUMULO-2579 Fix a bug where close was being written to replication table prematurely
A minc does not equate to a WAL no longer being used. The data up to the present length
in the WAL is ready to be replicated, but other tablets may still reference it. We can
only determine that a wal is "closed" when there are no additional references to it
in accumulo.metadata or when recovery occurs.