ACCUMULO-2575 Only remove status records when the entire file is replicated.
A problem arose when we inadvertently created a new status record as we were
in the middle of deleting it. Moved RemoveCompleteReplicationRecords from gc
into the master and run it as a part of the ReplicationDriver. This ensures
that we won't run concurrently with the StatusMaker which is the other
codepath which creates Status records. We can ensure that we get "isolation"
in our writes in this case.