ACCUMULO-2842 Allow configuration of FSDataOutputStream method used to sync WALs.
A critical piece of the bigtable design is using write-ahead logs to ensure
recovery after node failure. The append only file structure is meant to provide
a durable log of events so that the correct state can be reconstructed after
a failure when key-values have not been yet serialized to disk.

Hadoop 2 offers two levels of durability in regards to writes to files in HDFS. HFlush
ensures that all others clients who try to read the file you just wrote data to will
see that new data that was written. HSync will ensure that the data you wrote for that
file is actually sent to the underlying disk. As such, HFlush does not provide durability
when the node crashes without that data being flushed to the drive, while HSync does. HSync
does, however, increase the amount of time needed for sync/flush to complete (as hitting disk
consistently is slower than not).