ACCUMULO-2789 Added parameter to control the number of cells within a row
When running the stress test, it is now possible to configure the writer to
send mutations for rows that contain a random amount of column updates. This is
controlled via the --min-row-width, --max-row-width and --row-width-seed
parameters.

I did some internal shuffling to make randomly generating numbers within a
range more convenient. See the new RandomWithinRange class for details.