ACCUMULO-2918 Pay the penalty of embedding ZK hierarchy manipulation to avoid creating an upgrade path.
In a fresh 1.7.0 instance, this isn't an issue because {{accumulo init}} will create the necessary
parent nodes in ZooKeeper. However, for upgrades from 1.6 or 1.7 installs that were created
before the replication code was merged, the tserver will error out because the parent ZK nodes
do not exist (and ZK will not automatically create them).

While it's not ideal to embed this logic in the server processes, it's the better alternative
to create an upgrade path, because one does not yet exist.