ACCUMULO-2913 If we compact the table(s) it's possible that the replication entry will be cleaned up.
For either table, if we compact them after writing the data, the tablet will remove its WAL entry.
When no tablets reference that WAL anymore, it will be closed by the GC. Thus, the master might
remove the closed entry, but out test case might not have been able to read the metadata table
to verify that the record was there.

By not compacting the tables, we can ensure that the record will definitely be there.