ACCUMULO-2849 Add scanner batch size and write delay configurability to mem stress test
This commit adds one new option each to the Scan and Write components of the memory
stress test. The --scan-batch-size option for Scan configures the batch size of the
scanner it uses, while the --write-delay option for Write allows for a delay between
each mutation added to its batch writer. The reader.sh and writer.sh scripts are
updated for easy configuration of these new options.