ACCUMULO-2942 Fixes ShardedTableDistributionFormatterTest on all JVMs
ShardedTableDistributionFormatterTest.testAggregate() was failing due
to inconsistency in order to stats that were returned in the getStats()
method. This is due to the fact that the structure that holds the stats is
a HashMap.

The test was changed so that it does not assert that elements exists
at certain location in results, but rather it now checks that elements
exist in the list. The test was also slightly broadened so that it
checks that tuples of (day, count) exist in list.