ACCUMULO-2985 Make sure the ES used to stop processes with a timeout is stopped itself.
To use a Callable to apply a timeout when trying to stop the accumulo processes
in MiniAccumuloCluster, an ExecutorService is required. A single-thread ES was added
as a part of ACCUMULO-2764, however, it was neglected to be stopped itself
which added a non-daemon'ized thread to the JVM. This non-daemonized thread
prevents the JVM from exiting cleanly.

After using the ES to stop the Accumulo processes with a timeout, we
need to be sure to stop the ES as well. Test with a Mocked ES ensures that
the ES is stopped before MAC.stop() returns.