ACCUMULO-2994 Add explicit prefixes for iterator scopes and remove unnecessary configuration parsing.
When reading the Accumulo configuration, all configuration entries for all scopes were read. This
is highly unnecessary as all that was done was to warn against invalid table properties. Instead,
we can use the provided IteratorScope to explicitly pull out just the data that is needed from
the configuration, to reduce this processing.