ACCUMULO-3013 Inline reflection for Counter increment within ContinuousVerify
In 1.6, the CounterUtils class in server no longer exists, in favor of the
inlined method that was done in this commit. This alleviates modification
of the continuous-ingest scripts to launch jobs.

The CounterUtils class is still referenced by another class in server, but
both were also removed in 1.6 so they are not modified at all.