ACCUMULO-3016 Remove glob and use explicit file name for examples-simple jar
We don't bundle javadoc or sources in the assemble anymore so there
is no reason to try to use the glob to exclude those. Also, the version
string is also not included in the artifacts which make it even more
straightforward.

Also removed two tabs on the following lines and replaced with the
correct spaces for indentation.