ACCUMULO-2694 Ensure that the newly created table is assigned before continuing.
A table can be successfully created, but its tablets not yet assigned. There
was a race condition in the test where getMasterMonitorInfo was called
before the tablets for the TEST_TABLE were created. This made it appear that
the table did not exist.

We can mitigate the race condition by trying to read from our TEST_TABLE
before returning from the setup method. This ensures that the tablet for
our table is online and we should get the expected state from the test.