ACCUMULO-3055 Guard against failure due to multiple invocations of MAC.stop
MAC has allowed stop to be invoked a few times, so we want to ensure that we
don't start throwing an exception on multiple calls. There are also a few
race conditions (user invoking stop(), shutdown hook invoking stop(),
user invoking start again after MAC was stopped, etc) in which
we want to guard against this concurrency.

Adds synchronized to start and stop to ensure invocations of these
methods are exclusive and uses the `executor` member track start/stop
state. Adds a unit test to ensure multiple starts is an error, multiple
stops is not, reduces inline logging configuration, and applies
redirection to both surefire and failsafe tests.