ACCUMULO-3129 Pass credential provider into client configuration and set non-empty truststore password.
Making a non-empty password for the truststore makes things a bit
more reliable to actually work (I think empty passwords are disallowed
by the CredentialProvider impls). If the client is also using a
CredentialProvider, the ClientConfiguration also needs this property.
Adding the extra value to the enum passes it from accumulo-site.xml
into the generate client.conf by AbstractMacIT.