ACCUMULO-3150 Set 0 for monitor log4j port config value
The rest of the code is already in place to support
a randomly chosen port for the log4j forwarding to
the monitor. We just need to configure MAC to do this.

Added a test to ensure that ports we want to have
a value of '0' are in fact configured as such.