ACCUMULO-3182 Ensures that a WAL with a partial header (or completely) empty can be successfully recovered.
In the case where a WAL is observed which has a partial or missing header,
the file is treated as "empty": no log entries will be attempted to be
read and replayed from it. This should restore the normaly functionality
that was observed with 1.5 WALs.

Integration test was introduced which tests that the introduction of
either an empty WAL or a WAL with a partial header does not block
recovery of the tablet which references the "bad" WAL.

Additionally, an enum was introduced to remove the use of hard-coded
WAL recovery status marker files: "finished" and "failed".