ACCUMULO-3218 Set the list separator to \0 to 'disable' list interpretation
ZooKeeperInstance, because of how commons-configuration was parsing the
values, was treating a comma separated list of ZK servers (which
we intend to be a single value) as multiple and only returning
the first when getString was called on its key. We have no case
where we actually want values to be automatically parsed into an
array, so we can set it to a separator highly unlikely to be used.