ACCUMULO-3212 Create a copy from CachedConfiguration instead of a brand new Configuration
Quick/dirty test shows that creating a Configuration from an already initialized
Configuration object is much faster than a branch new Configuration. This ensures
that the created Configuration still has the necessary DFS properties and extracting
credential providers from HDFS works as expected with the default filesystem.

Added a test with MiniDFS and added an extra utility method to make things cleaner.
Also introduced the use of JUnit's Assume in the test case to check for the Hadoop
2.6.0 dependency on the classpath.