ACCUMULO-3355 Catch the case when we find no traces and quit early.
If we don't quit early, we end up creating this absurd date and sending
it back to the user. Also fixed two tests that were unnecessarily
nuking the entire trace table. The Scanner they create is bounded for
their trace session; there is no reason to delete the rest of the records.