ACCUMULO-3497 Quick check that the provided address uses the FQDN.
SASL relies heavily on DNS being properly configured. We can perform
a quick check for users that validates that the provided hostnames
in the "hosts" files in $ACCUMULO_CONF_DIR line up with the FQDN
as computed by InetAddress. This provides an intuitive message
as to the nature of their problem instead of a general GSS handshake
failure.