ACCUMULO-3556 Improve "scanner closed" error message.
When a client sees a "scanner closed" exception come out
of the BatchScanner impl, it is not intuitive that the
problem is within their code and not in Accumulo itself.
Provide a strong error message to inform the user that the
problem is in their code and not an Accumulo bug.