ACCUMULO-2462 Remove confusing indirect use of instance.dfs.uri
This removes a confusing indirect usage of the deprecated
instance.dfs.uri property, to resolve relative paths on the command-line
of PrintInfo. This removes the reliance on accumulo-site.xml for
determining which filesystem to use when not specifying a
fully-qualified file name.

There remains some reliance on accumulo-site.xml, however, due to
ACCUMULO-3566. It also doesn't remove the somewhat confusing behavior of
automatically looking at the local filesystem if the default hadoop
filesystem doesn't contain the file specified. To handle that special
case, I print out a message showing the actual file which it resolves
to.