ACCUMULO-3642 Create empty client configuration file to make unit tests pass.
If a client configuration exists on the current system, the Shell will
attempt to pick it up when one is not provided as an argument. This causes
spurious unit-test failures. These tests should be isolated from the
host environment and now set an empty client configuration as the
tests inherently expect.