ACCUMULO-3650 Push protobuf-java dependency into server-base.
This helps prevent protobuf from polluting the client's classpath.
Because Protobuf-generated code needs to be compiled with the version
included on the classpath, this helps prevent Accumulo from being
a pain point for clients (we add nothing more, just what hadoop bundles).

Needed to add a new RPC call to the master to support the implementation
of ReplicationOperations.drain(). Also removed a bunch of empty javadoc
class header blocks.