ACCUMULO-2325 Attempt to make example constraints more efficient.
Partially applied patch from Vikram Srivastava. Due to the nature
of constraints, it is far more likely that it will never fire. As
such, it makes sense that we optimize for that case. Avoiding the
allocation of a new collection (as the code already did) is far
better than always allocating a new list/set for every invocation
of the constraint. Tests were directly applied from the original patch.