ACCUMULO-1680 Added deleteauths shell command to reciprocate the addauths command
The ticket leaves an option for two commands.  The addauths commands was already added.
The ticket also suggests a transaction add/remove command using an atomic operation.
This was not done.  Not sure it is necessary.  It involves adjusting the API for the security operations and adding
more complex code in the zookeeper interaction.

Signed-off-by: Josh Elser <elserj@apache.org>