ACCUMULO-3952 Verify the log4j jar was found in the Hadoop installation.
The `find` command may fail (personally, ran into this when the cwd
was a directory that was not readable by the user). We miss this
and blindly proceed. Presently, Accumulo processes will fail miserably
when log4j isn't present. Since the user _may_ add a log4j jar to the
CLASSPATH, we avoid erroring out when CLASSPATH is non-empty