ACCUMULO-4060 Handle uncaught exceptions in FATE repo runner threads.
Any RuntimeException propagated from the internals of the FATE
repo-runner threads will caught the thread to exit. This creates
a situation where the Master can be running but be unresponsive
to any pending FATE operations like creating a table.

Make sure that these exceptions are caught and suppressed to ensure
that FATE ops can always be handled while the master is running.