ACCUMULO-4065 Create a separate connection cache for oneway Thrift calls.
Oneway Thrift calls return to the caller before the response from a server
is received. Even though all oneway calls are void methods, there is still
an implicit response sent by the server so that the client knows when
the RPC has completed. Because of this potential delay, a connection that
is re-used by a non-oneway call could see the response from the oneway
call, not its own (void or typed response message). We must separate
connections to ensure that this is avoided.